default light 256c : DarkYellow statements
In the current 256c version of the default light colorscheme, basic syntax
highlighting overwhelmingly relies on the bottom 8 ansi colors provided by the
terminal emulator.
This is a good thing, as this version of the colorscheme is also what vim falls
back to when it cannot identify whether the user's terminal background is light
or dark, as often happens through ssh.
Vim therefore relies on the assumption that the terminal palette provided by
the user suits their taste, is legible, or that the user will apply spot fixes
wherever needed.

Using brown is a hack that I believe was introduced to alleviate the fact that
the ansi3 yellow provided by xterm is barely legible on a pure white
background.

It is not vim's job to fix deficiencies in terminal emulators, no matter how
common they are, or were at the time this choice was made.